Extract ConsoleTestState
